pdf-rag-pipeline/requirements.txt
fastapi
uvicorn
pymongo[srv]
pymupdf
openai
google-generativeai
python-dotenv
PyJWT[crypto]
pydantic
pytest
